Only link edward component dynamically
Instead of providing both a static and dynamic library for this
component. This cuts compilation time by 50%. While I would prefer
static linkage by default CHICKEN does presently not rebuild the program
component of the egg (i.e. bin/edward) upon changes of the static
library and this is thus impractical for a development setup.