buffer: require creation of 'snapshots' for undo mechanism
Presently, these snapshots are created on a per editor operation
basis in lib/editor.scm. However, in the future it should be possible
to create these snapshots on a per ed command basis. Thereby easing
the implementation of the undo operation for the global command.

Since the undo stack is cleared periodically now this should also
prevent it from leaking memory.