TODO.md: Remove To-Do regarding multiline substitutes in global cmd
From POSIX.1-2008:

	A line can be split by substituting a <newline> into it. […]
	Such substitution cannot be done as part of a g or v command
	list.

So while GNU ed and BSD ed support this, it seems that POSIX explicitly
disallows multiline substitutions in g/v commands. For this reason,
edward is presently compliant to POSIX in this regard. On a side note,
edward presently does allow using multiline substitutions with the G/V
command.