Fix usage of substitute command with inserted newlines
Without this commit, regex-replace would simply return a string
containing a newline. This string would then be inserted into the
text editor buffer. However, the underlying assumption of the text
editor buffer is that each line is its own list element. This would
break stuff like usage of the substitute command with a suffixed print
command when a newline was inserted.

This commit fixes the issue by splitting the string returned by
regex-replace.