Align handling of substitute in global command with GNU/BSD ed
POSIX states the following regarding the substitute command:

	It is an error if the substitution fails on every addressed line.

The interesting question is how the term "addressed line" is supposed to
be interpreted in the context of the global command. GNU and BSD ed seem
to take the stance that, in the case of global command "addressed lines"
refers to the range of the global command. However, even if the
substitute command fails for every line of the global command they still
don't raise an error. I think this interaction is a bit underspecified
in POSIX for now just implement the behavior of GNU/BSD ed.