Don't use raise in io exception handler
Without this commit, the raise would cause the editor exception handler
to be invoked which would then invoke the io exception handler via raise
again. It is presently unclear to me whether this is a bug in the R7RS
CHICKEN implementation or intended behavior. In either case, it's
probably best also not error out on errors which are neither read nor
file errors. For example, shell syntax errors caused by popen(3).