Allow multiple addresses to be passed to commands
This is an explicit requirement of POSIX.1-2008:

	Any number of addresses can be provided to commands taking
	addresses; for example, "1,2,3,4,5p" prints lines 4 and 5,
	because two is the greatest valid number of addresses accepted
	by the print command.

This commit implements this requirement in edward. This required several
core changes to the edward code base. Most notably, the parser for
addresses has been re-worked and now returns a list of addresses. This
list can be converted to a pair using the addrlst->lpair procedure.
This procedure returns a pair of line numbers as a fixnum. As such, the
implemented editor commands now also operate on a fixnum.