Add more tests for zero command
From POSIX.1-2008:

	For consistency with the a and r commands and better user
	functionality, the i and c commands must also accept an address
	of 0, in which case 0i is treated as 1i and likewise for the c
	command.

However, as per <https://austingroupbugs.net/view.php?id=1130> the zero
address is not valid for the change command. As such, the zero address
only needs to be supported for the a, r, and i command. Furthermore, the
move commands excepts a zero addresses as a third addresses. The mark
command does not except a zero address as a third address. Both cases
are already covered by existing test cases.

Also update TODO.md accordingly.