ed: Don't separate address expansion from addr evaluation code
As it turns out, the POSIX ed specification does not allow application
of the omission rules prior to evaluation of the entire address chain.
For example, the address "7,5," must be evaluated as follows:

	7,5, [discard] -> 5, [omission] -> 5,5

While previously, edward would attempt to apply the omission rules first
(by grouping addresses to separators from the right) and then discard
address which is false. This can lead to incorrect final address
values for address chains which use ";" (e.g. ";,1,;").

Since evaluating address chains which use ";" requires access to the
editor object the entire address chain expansion code had to be moved
from addr.scm to editor.scm. The former still provides the procedure to
expand a single address though.