parse: allow reading past eof
The ed specification requires reading past EOF in order to treat
EOF analog to the quit command in command mode. Similarly, EOF
can be used to quit input mode. For both use-cases, it is required
to read past EOF.

Unfortunately, Scheme's read-char procedure is explicitly specified
in a way that forbids reading past EOF. For this reason, we need to
use the POSIX file API provided by CHICKEN in order to do so.
Unfortunately, we also need to retain compatibility with the old
read-char API in order to continue use of the parser with textual
input ports created via the open-input-string procedure.