Explicit handling of EOF for interactive G and V command
This edge case does not seem to be covered by POSIX. In the POSIX
nomenclature it is unclear whether the G and V command should be treated
as command- or input-mode in regards to the end-of-file indicator.

GNU ed raises an "unexpected end-of-file" when EOF is encountered
during these commands. This commit implements this behavior for edward.
Previously, edward would terminate entirely when an EOF was encountered
during parsing of these commands.