ed: Refactor handling of print commands
Some commands can be suffixed with print commands. Previously, print
commands which can be used as a suffix were hardcoded in
parse-print-cmd. As such, it was not possible to define custom print
suffix commands via the library interface. This commit modifies
cmd.scm to ensure that any command define via define-print-cmd can
be used as a suffix command.