Avoid use of match macro whenever reasonably possible
Match expansion does not seem to be subject to optimization. As it turns
out, case on integer constants is in fact compiled as a jump table by
csc(1) while match isn't.  So for primitive cases where only a
comparison against constant integers is performed it is worthwhile to
use case over match.

This can be debugged with `csc -debug 7 -O3`.

See also: The ##core#switch comment in optimizer.scm from chicken-core.