posix: in substitute only touch buffered lines that have been modified
This fixes a bug where marks would be invalidated if a mark was
included in the range of lines touched by the substitute as the
substitute command would re-write all lines (even if they were
unchanged!). This commit fixes this issue through multiple
editor-replace! invocations instead of performing a single
editor-replace! invocation on the entire range.

A test for this corner case is included. Furthermore, an additional
test case was added for ensure that marks _are_ invalided if the
marked line is modified by the substitute command.