replace: fix multiple insertions of matched string in substitute
Without this commit, edward would incorrectly handle multiple
substitutions when an ampersand was used in the replacement string.
This was due to the fact that the matched string (as indicated by the
ampersand in the replacement) was determined relative to the original
bytevector (instead of the modified bytevector). That is, the index
values returned for the matched string were incorrect if a previous
substitution modified the bytevector length.