replace: Indicate whether any replacements were performed
This is useful for the implementation of the substitute command.
According to the POSIX ed(1) specification:

	A substitution shall be considered to have been performed even
	if the replacement string is identical to the string that it
	replaces.

Prior to this commit, edward relied on string comparison to identify
performed replacement. This approach is not compatible with POSIX
and is fixed in this commit by having regex-replace indicate performed
modifications via its return value.

The regex-replace implement should be cleaned up in a future commit.