editor: Also reset line number in editor-reset!
This fixes a bug where an E command would not reset the line number
if a new file was created using it. For example:

	a
	foo
	bar
	.
	E new-file
	.=

Would incorrectly cause line number to be set to 2 even though it should
have been reset to zero after the E command. Unfortunately, it is
difficult to add a proper test case for this as the integration tests
will (correctly) terminate once the E command emits an error as the file
doesn't exist. The interactive tests, on the other hand, currently do
not handle commands with space correctly and will thus not be able to
execute the E command correctly. If the file does exists, then this
doesn't fail as the line number will then be (correctly) set to the last
inserted line.