GNUmakefile: Remove STATUS files when vendoring dependencies
The STATUS files contain information obtained from the CHICKEN
implementation fetching the vendored dependencies. This includes the
build-id (which refers to a Git commit in the CHICKEN repository).
Unfortunately, this build ID is only set if Git was installed when
CHICKEN was build which is not the case for the Alpine Linux CHICKEN
package, for example. Therefore, the STATUS file generated on an Alpine
machine is incompatible with a Debian CHICKEN installation.

AFAIK, the STATUS files should not be needed for vendored source code
hence we can safely remove them thus ensuring that the seem tarball
can be build on Alpine and Debian.