editor: Fix discarding of addrs for cmds expecting a single addr
Previously, the second address was discarded. That is, the first
was passed to the command. This is, however, incorrect as addresses
must be discarded starting at the beginning of an address list.

From POSIX:

	If more than the required number of addresses are provided
	to a command, the addresses specified first shall be
	evaluated and then discarded until the maximum number of
	valid addresses remain, for the specified command.

The addrlst->lpair function also correctly implements this algorithm, it
is just incorrectly implemented in the editor-xexec function. The
editor-exec function is fixed in this commit.