buffer: Use a vector instead of a list as the underlying data structure
This should allow for more efficient operations on the buffer (e.g.
vector-copy should be more efficient than sublist). Furthermore, this
allows using (chicken locative) in the future to create ed text marks
(thereby replacing the current editor-get-lnum hack).

For now, the API does not change in a backwards incompatible way and
it is still possible to add text (represented as a list) to the buffer.