Write files line-wise
Instead of joining the list of lines into a single string and writing
that at once using `write-string`. This is much more efficient for large
files as otherwise the majority of runtime is spent joining the list
to a string.

This was discovered while editing large files in Guix (e.g. the package
definitions from gnu/packages/python-xyz.scm) using the edward text
editor.