parse-with-context: retain failure continuation
Otherwise, parsers calling parse-with-context will discard the failure
continuation. This is because call-with-parse has a side-effect where it
modifies the failure continuation on the source via
`parse-stream-fk-set!`. Therefore, we shouldn't pass a custom fk here.
If we want a custom error message we could use parse-with-failure-reason
or similar here.

Test case, which I cannot commit right now because GNU ed actually
supports backslash substitution in the global command:

	g/foo/s/foo/test\
	foo\
	bar
	Q

POSIX does however not support such substitutions:

> A line can be split by substituting a <newline> into it. […]
> Such substitution cannot be done as part of a g or v command list.